processing-engine/requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
opencv-python-headless==4.8.0.74
Pillow==10.0.0
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0
